Added Health for Player with Dynamic UI for Damage
Added a TakeDamage function to Health class that removes X amount of health from the owner.
Damage can be set from EnemyController class.

Added UI Canvas to Player Prefab, with a label for Health and a dynamic label when damage is recieved by Player.
This canvas uses DisplayDamageUI script to display a moving label that showcases damage.

Updated Player Spawn, Death and EnemyCollision scripts to use the updated Health script.