Added Shooting Mechanic (Press 'Enter' to Fire)
Created a Bullet Prefab that moves in a specified direction and speed.
It also has a damage attribute that is currently set direcly by the Player Controller.
Bullet deals damage to Enemies, ignores Player collision and destroys itself when outside of screen.

Player Prefab has a BulletSpawner script that spawns the actual bullet to the screen.
Updated Health script to be more generic so that Enemy can use it as well.

You can fire a bullet by pressing the 'Enter' key.