Q1 WIP: Changed Player Health Damage
I refactored how Health.Decrement() worked to take in a damage input that defaults to 1. This meant I had to refactor less of the code for areas where Decrement was already used.
I then just had to add a Serialized public property (could have also used the old school private field + property style) for Damage to EnemyController and change the Player prefab to start at a health of 3.
Finally, I implemented the bulk of the logic in PlayerEnemyCollision. Curiously, though all the events are being called correctly (tested with console logs), the Player refuses to die, even when at 0 health. I've pinpointed it to this code section, but believe it might be related to Question 2 as well, so will return to this once that question is finished