Q1. Finished
Sorry for the large number of apparent changes. I implemented a visual text indicator using TMPro which meant I had to import the essentials for it to function.
I fixed the bug from the previous commit, which was caused by an extraneous if-statement in the PlayerDeath event. I also took a longer time to implement the damage text in a satisfactory way. Now, the player health can be assigned in the inspector, and when damage is taken, it follows the same event style as other gameplay functions.
Additionally, I also found and fixed an extra bug where I had to modify the PlayerSpawn event. After my initial changes, the player's health was set to 1 after respawning instead of refilling completely. That has been updated.