Q2. Added Immunity
The immunity was added by having a public boolean that can be read by PlayerEnemyCollision. PlayerEnemyCollision will only trigger player damage if the bool variable is false.
In Health, a Coroutine is triggered when damage is taken, and the bool is switched to true at the start of the coroutine, and then set to false after 1 second has elapsed.