Q3. Added Bullets
This was a bit tricky for me to implement while following the existing patterns, so took a while longer than I expected, with me having to refactor it several times.
Health scripts had to be added to the enemy prefabs, and the way the Health script triggered events also had to be adjusted.
Bullets fire in the direction the player is facing, and are pooled to reduce the amount of objects flying around and stressing the CPU.
There still exists a small bug after enabling the "hurt" animation on the enemies. If you bounce and land on them without moving, the colliders get stuck together and you collectively float upwards. I spent an extra 20ish minutes trying to resolve this, but decided to leave it in the interest of time