Q4:
This took about an hour to iron out the small bugs and refactoring previously written scripts to take advantage of the game modes. The easiest method I could see for having different game states was by using a Scriptable Object, Prefabs were an inelegant option that would be heavier on the system, and Addressables are more complex and overkill for the current goal.

A new component, GameModeManager, was added to the GameController object in the scene; in this component, the designer can simply drag and drop new game modes that follow the GameMode ScriptableObject pattern. The designer can also create these game modes through the asset menu and adjust the values in the inspector. There are also two different ways of storing position data for this game. The start position would ask the designer to manually input the vector values, while the spawn position works off of prefabs that can be dragged and dropped into the scriptable object. The second method would probably be easier for a designer to work with, but also means storing extra prefabs in the project.

Three sample Game Modes can be found in Assets > Scriptable Objects > Game Modes, and they can be tested by dragging them into the GameModeManager component on the GameController object in the sample scene.