Question 1 - Adjusted the Damage and Health code, and added feedback
I adjusted the PlayerDeath.cs file by getting rid of the if check on line 20 since the player was going to die regardless of the check.

Increased player's HP to 50

Adjusted the decrement function in Health.cs in order to inflict damage to the player between 1 and 7 randomly instead of decaying the player's HP.

Finally, I added a DisplayHealth function in Health.cs to provide feedback on the player's health and the damage inflicted on the player by an enemy in the console.