Question 2 - One second of invincibility after a collison with an enemy
I adjusted the Health.cs file by adding an InvincibilityFrames function that gives the player one second of immunity after colliding with an enemy. It is called in the Decrement function.

Added a second Decrement function for enemies that does not give them invincibility.

Added an isInvincible check in PlayerEnemyCollision.cs to prevent the player from taking damage in the event of a collision.

Fixed player health being decremented.