Question 3 - Gun weapon added
I added the Bullets.cs file into the mechanics folder and it defines how a bullet should work when fired in the game.

Added 8-bit sprites for the gun and bullets I found online.
Ref: https://tamamangrosse.itch.io/gun-sprite
https://www.artstation.com/artwork/Lev02P

Reworked PlayerController.cs to constantly be checking if the space bar was hit as that is the button to fire a bullet from the gun.

Added the FireGun function in PlayerController.cs to trigger when a bullet makes contact with an enemy.